+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.5.0]
+
 ### Added
 
 -   Added a `WithHistory` abstract model for integrating `django-simple-history` `HistoricalRecords`.
@@ -91,10 +93,11 @@ Initial release!
 
 -   Josh Thomas <josh@joshthomas.dev> (maintainer)
 
-[unreleased]: https://github.com/westerveltco/django-twc-toolbox/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/westerveltco/django-twc-toolbox/compare/v0.5.0...HEAD
 [0.2.1]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.2.1
 [0.2.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.2.0
 [0.1.1]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.1.1
 [0.3.0]: https://github.com/westerveltco/django-twc-toolbox.git/releases/tag/v0.3.0
 [0.3.1]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.3.1
 [0.4.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.4.0
+[0.5.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.5.0